// apps/main/src/models/v2/controlplane/ruleSummary.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { ModuleId } from './moduleId';

/**
 * A summary of a rule that matches resource requests against a module
 */
export interface RuleSummary {
  /** The Organization ID. */
  org_id: string;
  /** The unique identifier for the rule under the organization. */
  id: string;
  module_id: ModuleId;
  /** The time this rule was created */
  created_at: string;
  /** The resource type matched by this rule */
  resource_type: string;
  /** The resource class to match */
  resource_class: string;
  /** The resource id to match */
  resource_id?: string;
  /** The optional project id that this rule matches. */
  project_id?: string;
  /** The optional env type id that this rule matches. */
  env_type_id?: string;
  /** The optional environment id that this rule matches. */
  env_id?: string;
}